backend/requirements.txt
# FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# CORS and middleware (built into FastAPI)
# No additional CORS package needed

# OpenAI integration
openai==1.3.7

# File handling and storage
aiofiles==23.2.0
Pillow==10.4.0

# Data validation and serialization
pydantic==2.11.9
pydantic-settings==2.10.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2